Require case bodies to be code rather than values
i.e. terms with bindings rather than arbitrary unknown functions. For
the purposes of this experiment I did this by just insisting in the
evaluator that they should be lambdas and then going straight into the
body.

Then in PIR codegen we eta-expand unknown functions in case args so it
works out. We then rely on optimization getting rid of that indirection,
which it does, thanks to the new UPLC optimizer.